pyproject.toml
[project]
name = "clothing-upload-bot"
version = "0.1.0"
description = "Bot used as part of Filoxen Research Facilities to automatically upload clothing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.126.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
